requirements.txt
absl-py==1.1.0
aniso8601==9.0.1
astunparse==1.6.3
blinker==1.4
cachelib==0.9.0
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Caching==2.0.0
Flask-Cors==3.0.10
Flask-HTTPAuth==4.7.0
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-WTF==1.0.1
flatbuffers==1.12
gast==0.4.0
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
gunicorn==20.1.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
marshmallow==3.16.0
mongoengine==0.24.1
mongomock==4.0.0
numpy==1.23.0
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.4.0
pymongo==4.1.1
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
sentinels==1.0.0
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.1
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wincertstore==0.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
